infra/requirements.txt
psycopg2-binary==2.9.9
Faker==25.2.0
python-dotenv==1.0.1
